feat: 로그아웃 버튼 UI 구현 및 로직 작성
