+++ front/app/(home)/queries/auth.ts
@@ -1,0 +1,7 @@
+import { logout } from '../api/auth';
+
+export const homeQueryOptions = {
+  logout: () => ({
+    mutationFn: logout,
+  }),
+};
