--- front/app/page.tsx
@@ -1,5 +1,0 @@
-import { css } from '../styled-system/css';
-
-export default function Home() {
-  return <div className={css({ color: 'red' })}>hello world</div>;
-}
